[Refactoring] GlobalScope 코드 ViewModel 로 이동
